dev/backend/requirements.txt
Flask
Flask-Cors
PyJWT
Flask-SQLAlchemy
Flask-Migrate
psycopg2-binary
